pyproject.toml
[project]
name = "powerbi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests",
    "python-dotenv",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.8.0",
    "openai>=2.24.0",
    "numpy",
    "streamlit>=1.40.0",
    "fastapi>=0.120.0",
    "uvicorn[standard]>=0.34.0",
    "langchain",
    "langchain-openai",
    "langchain-community",
    "langgraph",
    "ipykernel",
]
